refactor(pages): 拆分 _dialogs 为独立组件文件，解耦 Service 构造函数，消除 PageObject 重复逻辑
- 将 _dialogs.py 拆分为 pages/components/ 下的独立文件，Overlay 基类同步移入
- CaptchaHandler / RecordChecker 构造函数不再持有 PageObject，改为方法参数注入
- LoginPage.login() 直接接收 auto_captcha 参数，简化 captcha_solver 调用链
- SeatMapOverlay.selectSeat 引入两层查找：先按 ID 直查，失败后遍历匹配
- 移除 ReserveView 中与 Dialog/Overlay 重复的方法（selectSeat、getAvailableTimeOptions）
- AutoLibPages 拆分 __initPagesServices / __initPagesFlows
- 修复 RecordsView.MORE_BTN 选择器被错误 snake_case 化（more_btn → moreBtn）
